refactor: create a core DNS client
