+++ challenge/dns01/client_nameservers.go
@@ -11,12 +11,12 @@ import (
 
 // checkRecursiveNameserversPropagation queries each of the recursive nameservers for the expected TXT record.
 func (c *Client) checkRecursiveNameserversPropagation(ctx context.Context, fqdn, value string) (bool, error) {
-	return c.checkNameserversPropagationCustom(ctx, fqdn, value, c.recursiveNameservers, false)
+	return c.checkNameserversPropagationCustom(ctx, fqdn, value, c.core.GetRecursiveNameservers(), false)
 }
 
 // checkRecursiveNameserversPropagation queries each of the authoritative nameservers for the expected TXT record.
 func (c *Client) checkAuthoritativeNameserversPropagation(ctx context.Context, fqdn, value string) (bool, error) {
-	authoritativeNss, err := c.lookupAuthoritativeNameservers(ctx, fqdn)
+	authoritativeNss, err := c.core.LookupAuthoritativeNameservers(ctx, fqdn)
 	if err != nil {
 		return false, err
 	}
@@ -31,7 +31,7 @@ func (c *Client) checkNameserversPropagationCustom(ctx context.Context, fqdn, va
 			ns = net.JoinHostPort(ns, c.authoritativeNSPort)
 		}
 
-		r, err := c.sendQueryCustom(ctx, fqdn, dns.TypeTXT, []string{ns}, false)
+		r, err := c.core.SendQueryCustom(ctx, fqdn, dns.TypeTXT, []string{ns}, false)
 		if err != nil {
 			return false, err
 		}
@@ -62,31 +62,4 @@ func (c *Client) checkNameserversPropagationCustom(ctx context.Context, fqdn, va
 	}
 
 	return true, nil
-}
-
-// lookupAuthoritativeNameservers returns the authoritative nameservers for the given fqdn.
-func (c *Client) lookupAuthoritativeNameservers(ctx context.Context, fqdn string) ([]string, error) {
-	var authoritativeNss []string
-
-	zone, err := c.FindZoneByFqdn(ctx, fqdn)
-	if err != nil {
-		return nil, fmt.Errorf("could not find zone: %w", err)
-	}
-
-	r, err := c.sendQuery(ctx, zone, dns.TypeNS, true)
-	if err != nil {
-		return nil, fmt.Errorf("NS call failed: %w", err)
-	}
-
-	for _, rr := range r.Answer {
-		if ns, ok := rr.(*dns.NS); ok {
-			authoritativeNss = append(authoritativeNss, strings.ToLower(ns.Ns))
-		}
-	}
-
-	if len(authoritativeNss) > 0 {
-		return authoritativeNss, nil
-	}
-
-	return nil, fmt.Errorf("[zone=%s] could not determine authoritative nameservers", zone)
 }
