+++ challenge/internal/client_timeout_unix.go
@@ -1,0 +1,8 @@
+//go:build !windows
+
+package internal
+
+import "time"
+
+// dnsTimeout is used as the default DNS timeout on Unix-like systems.
+const dnsTimeout = 10 * time.Second
