+++ challenge/internal/client_timeout_windows.go
@@ -1,0 +1,8 @@
+//go:build windows
+
+package internal
+
+import "time"
+
+// dnsTimeout is used as the default DNS timeout on Windows.
+const dnsTimeout = 20 * time.Second
