+++ challenge/internal/client_zone_test.go
@@ -1,4 +1,4 @@
-package dns01
+package internal
 
 import (
 	"testing"
