--- challenge/internal/dns.go
@@ -1,15 +1,0 @@
-package internal
-
-import "github.com/miekg/dns"
-
-func CreateDNSMsg(fqdn string, rtype uint16, recursive bool) *dns.Msg {
-	m := new(dns.Msg)
-	m.SetQuestion(fqdn, rtype)
-	m.SetEdns0(4096, false)
-
-	if !recursive {
-		m.RecursionDesired = false
-	}
-
-	return m
-}
