+++ challenge/internal/mock_test.go
@@ -1,0 +1,12 @@
+package internal
+
+import (
+	"github.com/miekg/dns"
+)
+
+func fakeNS(name, ns string) *dns.NS {
+	return &dns.NS{
+		Hdr: dns.RR_Header{Name: name, Rrtype: dns.TypeNS, Class: dns.ClassINET, Ttl: 172800},
+		Ns:  ns,
+	}
+}
